# Query Optimization - whereIn versus whereExists