@JosephSilber 2020-01-17
Do you know why it's faster?