@xp-bar 2020-01-27
My understanding is that `EXISTS` is typically faster when the results of the subquery are very large (since it stops searching after finding one result) whereas `WHERE IN` is typically faster if the results of your subquery is small (since the subquery runs once, vs `EXISTS` which runs once per row).

For my use case, it makes sense that `WHERE IN` is faster (lots of users, few permissions) but this may not be the case for everyone