@JosephSilber 2020-03-30
Do you still plan on revisiting this?