@xp-bar 2020-04-08
My apologies - when I said I would pull and retest master, I pulled and retested master, and then didn't leave a comment saying I had done so

Since it's been a bit, I'll do it again sometime this week and get back to you.

Cheers!