@xp-bar 2020-10-20
Hey again. I don't know if I'll get a chance to revisit this anytime soon, given how long it's been.

I can close this PR, and re-open a new one with the changes if I ever get a chance to double back on this.